Refactor folder's structure and framework
