+++ src/mux/echo.ts
@@ -1,6 +1,4 @@
-import Mux, { IMuxRequest, IRequestData } from '../framework/mux';
-import { IResponseRecord } from '../framework/response';
-import { Validator } from '../framework/validator';
+import { Mux, IMuxRequest, IRequestData, IResponseRecord, Validator } from '../framework';
 
 Mux.post<any>(
   '/v1/echo',
