+++ src/validators/index.ts
@@ -1,6 +1,7 @@
 export * from './article';
 export * from './comment';
 export * from './constant';
+export * from './user';
 
 // Please take note other file will be private and only available
 // to internal of validator
