+++ src/validators/pagination.ts
@@ -1,4 +1,4 @@
-import { Validator } from '../framework/validator';
+import { Validator } from '../framework';
 import { DefaultPagination } from './constant';
 
 export const PaginationValidator = new Validator(
